Simplificado

+++ Shared.Functions/SharedFunctions.cs
@@ -1,19 +1,11 @@
 ﻿using OpenQA.Selenium;
-using OpenQA.Selenium.Support.UI;
 using System;
-using System.Text;
+using System.Collections.Generic;
+using System.Drawing;
 using System.IO;
 using System.Net.Mail;
-using System.Diagnostics;
+using System.Text;
 using System.Threading;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Converters;
-using System.Dynamic;
-using System.Configuration;
-using System.Drawing;
 
 namespace Shared.Functions
 {
@@ -38,8 +30,8 @@ namespace Shared.Functions
             string oldFailedTest = sourceDir + "Failed_" + testName;
 
             // Delete files
-            File.Delete(oldPassedTest);
-            File.Delete(oldFailedTest);
+            if (File.Exists(oldPassedTest)) File.Delete(oldPassedTest);
+            if (File.Exists(oldFailedTest)) File.Delete(oldFailedTest);
         }
 
         public void createTestLog(string testCasePath, string text = "Test Case Execution Steps")
